construct an native AFNI hrf specification for '3dDeconvolve' with the 'stim_times' argument.

construct an native AFNI hrf specification for '3dDeconvolve' with the 'stim_times' argument.

afni_hrf(
  ...,
  basis = c("spmg1", "block", "dmblock", "tent", "csplin", "poly", "sin", "gam",
    "spmg2", "spmg3", "wav"),
  onsets = NULL,
  durations = NULL,
  prefix = NULL,
  subset = NULL,
  nbasis = 1,
  contrasts = NULL,
  id = NULL,
  start = NULL,
  stop = NULL
)

Arguments

...
the variable names, all of which must be present in the enclosing environment (e.g. an event_model object)

basis
the impulse response function or the name of a pre-supplied function, one of: "gamma", "spmg1", "spmg2", "spmg3", "bspline", "gaussian".

onsets
optional onsets override. If missing, onsets will be taken from the event_model

durations
optional durations override. If missing, onsets will be taken from the event_model

prefix
a character string that is prepended to the variable names and used to identify the term. Can be used to disambiguate two hrf terms with the same variable(s) but different onsets or basis functions.

subset
an expression indicating the subset of 'onsets' to keep

nbasis
number of basis functions -- only used for hemodynamic response functions (e.g. bspline) that take a variable number of bases.

contrasts
one or more contrast_spec objects created with the contrast function. If multiple contrasts are required, then these should be wrapped in a list or contrast_set.

id
a unique character identifier used to refer to term, otherwise will be determined from variable names.

start
the start of the window for sin/poly/csplin models

stop
the stop time for sin/poly/csplin models