afni_lm

afni_lm

afni_lm(
  fmri_mod,
  dataset,
  working_dir = ".",
  polort = -1,
  jobs = 1,
  censor = NULL,
  options = list()
)

Arguments

fmri_mod
an fmri_model object

dataset
an fmri_dataset object

working_dir
the working directory

polort
the number of polynomial baseline regressors (default is to suppress 'polort')

jobs
the number of jobs to use with '3dDeconvolve'

censor
a list of censoring vectors, one per run. Or a single vector equal to total number of scans.

options
a list of options to be sent to 3dDeconvolve

Examples

etab <- data.frame(onset=c(1,30,15,25), fac=factor(c("A", "B", "A", "B")), run=c(1,1,2,2))
dset <- fmri_dataset(scans=c("s1.nii", "s2.nii"), mask="mask.nii", TR=1, run_length=c(50,50),event_table=etab)

emodel <- event_model(onset ~ hrf(fac), block = ~ run, data=etab, sampling_frame=dset$sampling_frame)

#> Warning: The `x` argument of `as_tibble.matrix()` must have unique column names if `.name_repair` is omitted as of tibble 2.0.0.
#> Using compatibility `.name_repair`.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

bmodel <- baseline_model("bs", degree=4, sframe=dset$sampling_frame)

#> [1] "baseline_bs_4 : matrix"

fmod <- fmri_model(emodel, bmodel)
alm <- afni_lm(fmod, dset, jobs=2, options=list(tout=TRUE))