construct a sampling_frame

A sampling_frame describes the block structure and temporal sampling of an fMRI paradigm.

sampling_frame(blocklens, TR, start_time = TR/2, precision = 0.1)

Arguments

blocklens

the number of scans in each block, a vector

TR

the repetition time in seconds; i.e. the spacing between consecutive image acquisitions.

start_time

the offset of first scan of each block (default is TR/2)

precision

the discrete sampling interval used for convolution with hemodynamic response function.

Examples

frame <- sampling_frame(blocklens=c(100,100, 100), TR=2, precision=.5)

## the relative time (with respect to the last block) in seconds of each sample/acquisition
sam <- samples(frame)
## the global time (with respect to first block) of each sample/acquisition
gsam <- samples(frame, global=TRUE)