estimate trialwise beta coefficients for a dataset

estimate trialwise beta coefficients for a dataset

estimate_betas(x, ...)

# S3 method for fmri_dataset
estimate_betas(
  x,
  fixed = NULL,
  ran,
  block,
  method = c("mixed", "pls", "pls_searchlight", "pls_global", "ols"),
  basemod = NULL,
  radius = 8,
  niter = 8,
  ncomp = 4,
  lambda = 0.01,
  ...
)

# S3 method for matrix_dataset
estimate_betas(
  x,
  fixed = NULL,
  ran,
  block,
  method = c("mixed", "pls", "pls_global", "ols"),
  basemod = NULL,
  ncomp = 4,
  lambda = 0.01,
  ...
)

# S3 method for latent_dataset
estimate_betas(
  x,
  fixed = NULL,
  ran,
  block,
  method = c("mixed", "pls", "pls_global", "ols"),
  basemod = NULL,
  ncomp = 4,
  lambda = 0.01,
  prewhiten = TRUE,
  ...
)

Arguments

x

the dataset

...

extra args

fixed

a formula for the fixed regressors that model constant effects (i.e. non-varying over trials)

ran

a formula for the random (trialwise) regressors that model single trial effects

block

a formula for the block factor

method

the regression method for estimating trialwise betas

basemod

baseline_model instance to regress out of data before beta estimation

radius

the radius in mm for pls_searchlight approach

niter

number of searchlight iterations for method "pls_searchlight"

ncomp

number of pls components for method "pls" and "pls_searchlight" and "pls_global"

lambda

lambda parameter (not currently used)

prewhiten

whether to prewhiten basis set using auto.arima

Details

The method arguments allows for several beta estimation approaches

mixed uses a linear mixed effects modeling of trialwise random effects as implemented in the rrBLUP package.

pls uses separate partial least squares for each voxel to estimate trialwise betas

pls_searchlight estimates pls solutions over searchlight windows and averages the beta estimates

pls_global estimates a single multiresponse pls solution, where the Y matrix is the full data matrix.

ols ordinary least squares estimate of betas -- no regularization

Examples

facedes <- read.table(system.file("extdata", "face_design.txt", package = "fmrireg"), header=TRUE)
facedes$frun <- factor(facedes$run)
scans <- paste0("rscan0", 1:6, ".nii")

dset <- fmri_dataset(scans=scans,mask="mask.nii",TR=1.5,run_length=rep(436,6),event_table=facedes)
fixed = onset ~ hrf(run)
ran = onset ~ trialwise()
block = ~ run