evaluate a function over a sampling grid

given an object to be evaluated and an input sample "grid", evaluate the object.

evalute a regressor function over a sampling grid

evaluate(x, grid, ...)

# S3 method for HRF
evaluate(
  x,
  grid,
  amplitude = 1,
  duration = 0,
  precision = 0.2,
  summate = TRUE,
  normalize = FALSE,
  ...
)

# S3 method for regressor
evaluate(x, grid, precision = 0.33, use_conv = FALSE, ...)

Arguments

x

the object to evaluate

grid

the sampling grid. A vector of real values in seconds.

...

extra args

amplitude

the scaling value for the event

duration

the duration of the event

precision

the sampling precision for the hrf. This parameter is passed to evaluate.HRF

summate

whether the HRF response increases its amplitude as a function of stimulus duration.

normalize

scale output so that peak is 1.

use_conv

use fast convolution approach

Examples

hrf1 <- evaluate(HRF_SPMG1, grid=seq(0,20,by=1.5), duration=2, precision=.1)

# the same, except now turn off temporal summation.
hrf2 <- evaluate(HRF_SPMG1, grid=seq(0,20,by=1.5), duration=2, precision=.1,summate=FALSE)

frame <- sampling_frame(blocklens=100, TR=2)
reg <- regressor(onsets=c(10,20,35, 47,52, 68, 79,86), hrf=HRF_SPMG1)
evaluate(reg, samples(frame))
#>   [1]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>   [6]  3.064436e-03  1.007784e-01  1.753710e-01  1.271138e-01  5.746414e-02
#>  [11]  1.657867e-02  9.302341e-02  1.602333e-01  1.125000e-01  4.673899e-02
#>  [16]  6.956570e-03 -1.123682e-02 -1.678499e-02  2.075367e-02  1.452236e-01
#>  [21]  1.537497e-01  8.654527e-02  3.037279e-02 -4.075183e-05  2.303229e-02
#>  [26]  1.405722e-01  1.505830e-01  1.822762e-01  2.025447e-01  1.253565e-01
#>  [31]  4.361059e-02 -2.488198e-03 -2.078230e-02 -2.375180e-02 -1.642433e-02
#>  [36]  8.762009e-02  1.677196e-01  1.231828e-01  5.564567e-02  1.274578e-02
#>  [41]  2.801979e-02  1.409813e-01  1.457589e-01  8.239033e-02  1.262489e-01
#>  [46]  1.725584e-01  1.127046e-01  4.120336e-02  3.783271e-04 -1.641113e-02
#>  [51] -2.002767e-02 -1.705209e-02 -1.182045e-02 -7.007883e-03 -3.652946e-03
#>  [56] -1.708338e-03 -7.279723e-04 -2.862478e-04 -1.049366e-04 -3.617087e-05
#>  [61] -1.164267e-05 -3.626379e-06 -1.081010e-06  0.000000e+00  0.000000e+00
#>  [66]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [71]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [76]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [81]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [86]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [91]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [96]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00