evaluate a function over a sampling grid

given an object to be evaluated and a an input sample ("grid"), evaluate the object.

evalute a regressor function over a sampling grid

evaluate(x, grid, ...)

# S3 method for HRF
evaluate(x, grid, amplitude = 1, duration = 0,
  precision = 0.2, summate = TRUE, normalize = FALSE)

# S3 method for regressor
evaluate(x, grid, precision = 0.2)

Arguments

x
the object to evaluate

grid
the sampling grid

...
extra args

amplitude
the scaling value for the event

duration
the duration of the event

precision
the temporal resolution used for computing summed responses when duration > 0

summate
whether the HRF response increases its amplitude as a function of stimulus duration.

normalize
scale output so that peak is 1.

grid
the sampling grid. A vector of real values in seconds.

precision
the sampling precision for the hrf. This parameter is passed to evaluate.HRF

Examples

hrf1 <- evaluate(HRF_SPMG1, grid=seq(0,20,by=1.5), duration=2, precision=1)

# the same, except now turn off temporal summation.
hrf2 <- evaluate(HRF_SPMG1, grid=seq(0,20,by=1.5), duration=2, precision=1,summate=FALSE)

frame <- sampling_frame(blocklens=100, TR=2)
reg <- regressor(onsets=c(10,20), hrf=HRF_SPMG1)
evaluate(reg, samples(frame))

#>   [1]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.01747402
#>   [7]  0.57465835  1.00000000  0.72482810  0.32767192  0.09453487  0.53043788
#>  [13]  0.91368226  0.64149721  0.26651499  0.03966774 -0.06407456 -0.08631774
#>  [19] -0.08333089 -0.06115694 -0.03739311 -0.01985409  0.00000000  0.00000000
#>  [25]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [31]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [37]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [43]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [49]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [55]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [61]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [67]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [73]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [79]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [85]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [91]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [97]  0.00000000  0.00000000  0.00000000  0.00000000